scripts/buildkite/pyproject.toml
[tool.poetry]
name = "buildkite_wiring"
version = "0.1.0"
description = ""
authors = ["Jeremiah Campbell <jeremiah.campbell@annalect.com>"]

[tool.poetry.dependencies]
python = "^3.7"
sh = "^1.12"
PyGithub = "^1.43"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"